# server/src/public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Oya Browser API",
    "description": "REST API for controlling browsers connected to the Oya Browser server. Browsers connect via WebSocket; AI tools and clients use this REST API (or MCP) to send commands.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://browser.getoya.ai",
      "description": "Production"
    },
    {
      "url": "http://localhost:3100",
      "description": "Local development"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/api/control": {
      "get": {
        "summary": "Project operational overview",
        "operationId": "getApiControl",
        "tags": [
          "Durable control"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or revoked credential"
          },
          "403": {
            "description": "Insufficient role"
          },
          "409": {
            "description": "Lifecycle, idempotency, or control lease conflict"
          },
          "503": {
            "description": "Storage or session owner unavailable"
          }
        }
      }
    },
...
    "/api/control/sessions/{id}/control": {
      "post": {
        "summary": "Transfer human control; release leaves agent paused",
        "operationId": "postApiControlSessionsIdControl",
        "tags": [
          "Durable control"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or revoked credential"
          },
          "403": {
            "description": "Insufficient role"
          },
          "409": {
            "description": "Lifecycle, idempotency, or control lease conflict"
          },
          "503": {
            "description": "Storage or session owner unavailable"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "action": {
                    "type": "string",
                    "enum": [
                      "acquire",
                      "release",
                      "resume"
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/control/project": {
      "patch": {
        "summary": "Update limits, rates, retention and admission policy",
        "operationId": "patchApiControlProject",
        "tags": [
          "Durable control"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or revoked credential"
          },
          "403": {
            "description": "Insufficient role"
          },
          "409": {
            "description": "Lifecycle, idempotency, or control lease conflict"
          },
          "503": {
            "description": "Storage or session owner unavailable"
          }
        },
        "description": "Requires an administrator credential.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "maxConcurrent": {
                    "type": "integer",
                    "nullable": true
                  },
                  "budgetUsd": {
                    "type": "number",
                    "nullable": true
                  },
                  "rates": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "number",
                      "minimum": 0
                    }
                  },
                  "policy": {
                    "type": "object"
                  },
                  "recordingDays": {
                    "type": "integer",
                    "minimum": 1
                  },
                  "auditDays": {
                    "type": "integer",
                    "minimum": 1
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/control/members": {
      "get": {
        "summary": "List project members",
        "operationId": "getApiControlMembers",
        "tags": [
          "Durable control"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or revoked credential"
          },
          "403": {
            "description": "Insufficient role"
          },
          "409": {
            "description": "Lifecycle, idempotency, or control lease conflict"
          },
          "503": {
            "description": "Storage or session owner unavailable"
          }
        },
        "description": "Requires an administrator credential."
      }
    },
    "/api/control/members/invite": {
      "post": {
        "summary": "Create one-use seven-day invitation",
        "operationId": "postApiControlMembersInvite",
        "tags": [
          "Durable control"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or revoked credential"
          },
          "403": {
            "description": "Insufficient role"
          },
          "409": {
            "description": "Lifecycle, idempotency, or control lease conflict"
          },
          "503": {
            "description": "Storage or session owner unavailable"
          }
        },
        "description": "Requires an administrator credential.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "role": {
                    "type": "string",
                    "enum": [
                      "viewer",
                      "operator",
                      "administrator"
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/control/members/{id}": {
      "delete": {
        "summary": "Remove member and revoke linked credentials",
        "operationId": "deleteApiControlMembersId",
        "tags": [
          "Durable control"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or revoked credential"
          },
          "403": {
            "description": "Insufficient role"
          },
          "409": {
            "description": "Lifecycle, idempotency, or control lease conflict"
          },
          "503": {
            "description": "Storage or session owner unavailable"
          }
        },
        "description": "Requires an administrator credential.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/mcp/{browserId}": {
      "post": {
        "summary": "MCP Streamable HTTP endpoint",
        "description": "Model Context Protocol endpoint for AI tool integration (Claude Desktop, Cursor, etc.). Implements the MCP Streamable HTTP transport.",
        "operationId": "mcpPost",
        "parameters": [
          { "$ref": "#/components/parameters/browserId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "MCP JSON-RPC request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "MCP JSON-RPC response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "MCP JSON-RPC response"
                }
              }
            }
          },
          "404": {
            "description": "Browser not connected",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      },
      "get": {
        "summary": "MCP SSE endpoint",
        "operationId": "mcpGet",
        "parameters": [
          { "$ref": "#/components/parameters/browserId" }
        ],
        "responses": {
          "200": {
            "description": "MCP SSE stream",
            "content": {
              "text/event-stream": {
                "schema": { "type": "string" }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "MCP session cleanup",
        "operationId": "mcpDelete",
        "parameters": [
          { "$ref": "#/components/parameters/browserId" }
        ],
        "responses": {
          "200": {
            "description": "Session closed"
          }
        }
      }
    },
...
    "/mcp/pool": {
      "post": {
        "summary": "Pool MCP endpoint",
        "operationId": "mcpPool",
        "description": "MCP Streamable HTTP endpoint for the browser pool. Round-robins commands across all browsers sharing the caller's API key. navigate/analyze_page advance the round-robin; click/type/screenshot stay pinned to the last-used browser.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "MCP JSON-RPC request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "MCP JSON-RPC response or SSE stream",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "MCP JSON-RPC response"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key passed as Bearer token. Create one from the dashboard after signing in."
      }
    },
    "parameters": {
      "browserId": {
        "name": "browserId",
        "in": "path",
        "required": true,
        "schema": { "type": "string" },
        "description": "Browser session ID (from GET /browsers)"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        }
      },
...
    }
  }
}